refactor: create an initialize class for the tree table task
